refactor: :art: better naming and structure for discord api files
